Add files via upload
scatterplot,relplot, difference between figure object, axis object is covered. We use hue, style and size parameters to judge datapoints based on different columns. The use of col and row parameters(used in figure level functions), use of col-wrap. We view the use of distplot,histplot,kdeplot,rugplot. It can be used to make a bivariate displot(kind=hist) and also bivariate kde plot. It can be used to make a heatmap out of pivot table with the use of annot, linewidth, cmap parameter. It can also be used to make cluster map.